Accept floats

+++ db/migrate/20231216184339_create_metrics.rb
@@ -1,9 +1,9 @@
 class CreateMetrics < ActiveRecord::Migration[7.1]
   def change
     create_table :metrics do |t|
-      t.integer :weight
-      t.integer :fat_percentage
-      t.integer :muscle_percentage
+      t.float :weight
+      t.float :fat_percentage
+      t.float :muscle_percentage
       t.integer :user_id
 
       t.timestamps
